fixes test and doc
