+++ services/payments/tests/conftest.py
@@ -24,7 +24,7 @@ from pytest_simcore.helpers.rawdata_fakers import (
     random_user,
 )
 from pytest_simcore.helpers.typing_env import EnvVarsDict
-from pytest_simcore.helpers.utils_envs import setenvs_from_dict
+from pytest_simcore.helpers.utils_envs import load_dotenv, setenvs_from_dict
 from servicelib.utils_secrets import generate_token_secret_key
 from simcore_postgres_database.models.payments_transactions import (
     PaymentTransactionState,
@@ -73,6 +73,56 @@ def fake_user_name(faker: Faker) -> str:
 @pytest.fixture
 def fake_password(faker: Faker) -> str:
     return faker.password(length=10)
+
+
+def pytest_addoption(parser: pytest.Parser):
+    group = parser.getgroup(
+        "external_environment",
+        description="Replaces mocked services with real ones by passing actual environs and connecting directly to external services",
+    )
+    group.addoption(
+        "--external-envfile",
+        action="store",
+        type=Path,
+        default=None,
+        help="Path to an env file. Consider passing a link to repo configs, i.e. `ln -s /path/to/osparc-ops-config/repo.config`",
+    )
+    group.addoption(
+        "--external-email",
+        action="store",
+        type=str,
+        default=None,
+        help="An email for test_services_notifier_email",
+    )
+
+
+@pytest.fixture(scope="session")
+def external_environment(request: pytest.FixtureRequest) -> EnvVarsDict:
+    """
+    If a file under test folder prefixed with `.env-secret` is present,
+    then this fixture captures it.
+
+    This technique allows reusing the same tests to check against
+    external development/production servers
+    """
+    envs = {}
+    if envfile := request.config.getoption("--external-envfile"):
+        assert isinstance(envfile, Path)
+        print("🚨 EXTERNAL: external envs detected. Loading", envfile, "...")
+        envs = load_dotenv(envfile)
+        assert "PAYMENTS_GATEWAY_API_SECRET" in envs
+        assert "PAYMENTS_GATEWAY_URL" in envs
+
+    return envs
+
+
+@pytest.fixture
+def env_devel_dict(
+    env_devel_dict: EnvVarsDict, external_environment: EnvVarsDict
+) -> EnvVarsDict:
+    if external_environment:
+        return external_environment
+    return env_devel_dict
 
 
 @pytest.fixture
